+++ uberduck_ml_dev/models/tacotron2.py
@@ -1,692 +1,19 @@
-__all__ = [
-    "Decoder",
-    "Prenet",
-    "Postnet",
-    "Encoder",
-    "DEFAULTS",
-    "config",
-    "DEFAULTS",
-    "Tacotron2",
-]
-
 from torch import nn
-from .base import TTSModel
-from .common import Conv1d, LinearNorm
-from typing import Optional, List
+import numpy as np
+import torch
+from torch.nn import functional as F
 
 # NOTE (Sam): components is redundant with common - we needed a different name to make the package imports work as we migrate to a folder based approach
-from .components.attention import Attention
-from ..text.symbols import symbols
 from ..vendor.tfcompat.hparam import HParams
 from ..utils.utils import get_mask_from_lengths
 from ..data.batch import Batch
-
-import numpy as np
-import torch
-from torch.autograd import Variable
-from torch.cuda.amp import autocast
-from torch.nn import functional as F
-
-
-class Decoder(nn.Module):
-    def __init__(self, hparams):
-        super().__init__()
-        self.n_mel_channels = hparams.n_mel_channels
-        self.n_frames_per_step_initial = hparams.n_frames_per_step_initial
-        self.n_frames_per_step_current = hparams.n_frames_per_step_initial
-        self.encoder_embedding_dim = hparams.encoder_embedding_dim
-        self.attention_rnn_dim = hparams.attention_rnn_dim
-        self.decoder_rnn_dim = hparams.decoder_rnn_dim
-        self.prenet_dim = hparams.prenet_dim
-        self.max_decoder_steps = hparams.max_decoder_steps
-        self.gate_threshold = hparams.gate_threshold
-        self.p_attention_dropout = hparams.p_attention_dropout
-        self.p_decoder_dropout = hparams.p_decoder_dropout
-        self.p_teacher_forcing = hparams.p_teacher_forcing
-        self.cudnn_enabled = hparams.cudnn_enabled
-        self.attention_hidden = torch.tensor([])
-        self.attention_cell = torch.tensor([])
-        self.decoder_hidden = torch.tensor([])
-        self.decoder_cell = torch.tensor([])
-        self.attention_weights = torch.tensor([])
-        self.attention_weights_cum = torch.tensor([])
-        self.attention_context = torch.tensor([])
-        self.memory = torch.tensor([])
-        self.processed_memory = torch.tensor([])
-        self.mask = torch.tensor([])
-
-        self.prenet = Prenet(
-            hparams.n_mel_channels,
-            [hparams.prenet_dim, hparams.prenet_dim],
-        )
-
-        self.attention_rnn = nn.LSTMCell(
-            hparams.prenet_dim + self.encoder_embedding_dim,
-            hparams.attention_rnn_dim,
-        )
-
-        self.attention_layer = Attention(
-            hparams.attention_rnn_dim,
-            self.encoder_embedding_dim,
-            hparams.attention_dim,
-            hparams.attention_location_n_filters,
-            hparams.attention_location_kernel_size,
-            fp16_run=hparams.fp16_run,
-        )
-
-        self.decoder_rnn = nn.LSTMCell(
-            hparams.attention_rnn_dim + self.encoder_embedding_dim,
-            hparams.decoder_rnn_dim,
-            1,
-        )
-
-        self.linear_projection = LinearNorm(
-            hparams.decoder_rnn_dim + self.encoder_embedding_dim,
-            hparams.n_mel_channels * hparams.n_frames_per_step_initial,
-        )
-
-        self.gate_layer = LinearNorm(
-            hparams.decoder_rnn_dim + self.encoder_embedding_dim,
-            1,
-            bias=True,
-            w_init_gain="sigmoid",
-        )
-
-    def set_current_frames_per_step(self, n_frames: int):
-        self.n_frames_per_step_current = n_frames
-
-    def get_go_frame(self, memory):
-        """Gets all zeros frames to use as first decoder input
-        PARAMS
-        ------
-        memory: decoder outputs
-
-        RETURNS
-        -------
-        decoder_input: all zeros frames
-        """
-        B = memory.size(0)
-        decoder_input = memory.data.new_zeros(B, self.n_mel_channels)
-        return decoder_input
-
-    def initialize_decoder_states(self, memory, mask):
-        """Initializes attention rnn states, decoder rnn states, attention
-        weights, attention cumulative weights, attention context, stores memory
-        and stores processed memory
-        PARAMS
-        ------
-        memory: Encoder outputs
-        mask: Mask for padded data if training, expects None for inference
-        """
-        B = memory.size(0)
-        MAX_TIME = memory.size(1)
-
-        self.attention_hidden = memory.data.new_zeros(B, self.attention_rnn_dim)
-        self.attention_cell = memory.data.new_zeros(B, self.attention_rnn_dim)
-
-        self.decoder_hidden = memory.data.new_zeros(B, self.decoder_rnn_dim)
-        self.decoder_cell = memory.data.new_zeros(B, self.decoder_rnn_dim)
-        self.attention_weights = memory.data.new_zeros(B, MAX_TIME)
-        self.attention_weights_cum = memory.data.new_zeros(B, MAX_TIME)
-        self.attention_context = memory.data.new_zeros(B, self.encoder_embedding_dim)
-
-        self.memory = memory
-        self.processed_memory = self.attention_layer.memory_layer(memory)
-        self.mask = mask
-
-    def parse_decoder_inputs(self, decoder_inputs):
-        """Prepares decoder inputs, i.e. mel outputs
-        PARAMS
-        ------
-        decoder_inputs: inputs used for teacher-forced training, i.e. mel-specs
-
-        RETURNS
-        -------
-        inputs: processed decoder inputs
-
-        """
-        # (B, n_mel_channels, T_out) -> (B, T_out, n_mel_channels)
-
-        decoder_inputs = decoder_inputs.transpose(1, 2)
-        decoder_inputs = decoder_inputs.contiguous()
-        decoder_inputs = decoder_inputs.view(
-            decoder_inputs.size(0),
-            int(decoder_inputs.size(1) / self.n_frames_per_step_current),
-            -1,
-        )
-        # (B, T_out, n_mel_channels) -> (T_out, B, n_mel_channels)
-        decoder_inputs = decoder_inputs.transpose(0, 1)
-        return decoder_inputs
-
-    def parse_decoder_outputs(
-        self,
-        mel_outputs,
-        gate_outputs: List[torch.Tensor],
-        alignments: List[torch.Tensor],
-    ):
-        """Prepares decoder outputs for output
-        PARAMS
-        ------
-        mel_outputs:
-        gate_outputs: gate output energies
-        alignments:
-
-        RETURNS
-        -------
-        mel_outputs:
-        gate_outpust: gate output energies
-        alignments:
-        """
-        # (T_out, B) -> (B, T_out)
-        alignments = torch.stack(alignments).transpose(0, 1)
-        # (T_out, B) -> (B, T_out)
-        gate_outputs = torch.stack(gate_outputs)
-        if len(gate_outputs.size()) > 1:
-            gate_outputs = gate_outputs.transpose(0, 1)
-        else:
-            gate_outputs = gate_outputs[None]
-        gate_outputs = gate_outputs.contiguous()
-        # (B, T_out, n_mel_channels * n_frames_per_step) -> (B, T_out * n_frames_per_step, n_mel_channels)
-        mel_outputs = mel_outputs.view(mel_outputs.size(0), -1, self.n_mel_channels)
-        # (B, T_out, n_mel_channels) -> (B, n_mel_channels, T_out)
-        mel_outputs = mel_outputs.transpose(1, 2)
-
-        return mel_outputs, gate_outputs, alignments
-
-    def decode(self, decoder_input, attention_weights: Optional[torch.Tensor]):
-        """Decoder step using stored states, attention and memory
-        PARAMS
-        ------
-        decoder_input: previous mel output
-
-        RETURNS
-        -------
-        mel_output:
-        gate_output: gate output energies
-        attention_weights:
-        """
-        cell_input = torch.cat((decoder_input, self.attention_context), -1)
-        self.attention_hidden, self.attention_cell = self.attention_rnn(
-            cell_input, (self.attention_hidden, self.attention_cell)
-        )
-        self.attention_hidden = F.dropout(
-            self.attention_hidden, self.p_attention_dropout, self.training
-        )
-        self.attention_cell = F.dropout(
-            self.attention_cell, self.p_attention_dropout, self.training
-        )
-
-        attention_weights_cat = torch.cat(
-            (
-                self.attention_weights.unsqueeze(1),
-                self.attention_weights_cum.unsqueeze(1),
-            ),
-            dim=1,
-        )
-        self.attention_context, self.attention_weights = self.attention_layer(
-            self.attention_hidden,
-            self.memory,
-            self.processed_memory,
-            attention_weights_cat,
-            self.mask,
-            attention_weights,
-        )
-
-        self.attention_weights_cum += self.attention_weights
-        decoder_input = torch.cat((self.attention_hidden, self.attention_context), -1)
-        self.decoder_hidden, self.decoder_cell = self.decoder_rnn(
-            decoder_input, (self.decoder_hidden, self.decoder_cell)
-        )
-        self.decoder_hidden = F.dropout(
-            self.decoder_hidden, self.p_decoder_dropout, self.training
-        )
-        self.decoder_cell = F.dropout(
-            self.decoder_cell, self.p_decoder_dropout, self.training
-        )
-
-        decoder_hidden_attention_context = torch.cat(
-            (self.decoder_hidden, self.attention_context), dim=1
-        )
-
-        decoder_output = self.linear_projection(decoder_hidden_attention_context)
-
-        gate_prediction = self.gate_layer(decoder_hidden_attention_context)
-        return decoder_output, gate_prediction, self.attention_weights
-
-    def forward(self, memory, decoder_inputs, memory_lengths):
-        """Decoder forward pass for training
-        PARAMS
-        ------
-        memory: Encoder outputs
-        decoder_inputs: Decoder inputs for teacher forcing. i.e. mel-specs
-        memory_lengths: Encoder output lengths for attention masking.
-
-        RETURNS
-        -------
-        mel_outputs: mel outputs from the decoder
-        gate_outputs: gate outputs from the decoder
-        alignments: sequence of attention weights from the decoder
-        """
-        B = memory.size(0)
-        decoder_inputs = self.parse_decoder_inputs(decoder_inputs)
-        decoder_inputs = decoder_inputs.reshape(
-            -1, decoder_inputs.size(1), self.n_mel_channels
-        )
-        decoder_input = self.get_go_frame(memory).unsqueeze(0)
-        decoder_inputs = torch.cat((decoder_input, decoder_inputs), dim=0)
-        decoder_inputs = self.prenet(decoder_inputs)
-
-        self.initialize_decoder_states(
-            memory, mask=~get_mask_from_lengths(memory_lengths)
-        )
-
-        mel_outputs = torch.empty(
-            B, 0, self.n_frames_per_step_current * self.n_mel_channels
-        )
-        if self.cudnn_enabled:
-            mel_outputs = mel_outputs.cuda()
-        gate_outputs, alignments = [], []
-        desired_output_frames = decoder_inputs.size(0) / self.n_frames_per_step_current
-        while mel_outputs.size(1) < desired_output_frames - 1:
-            if (
-                mel_outputs.size(1) == 0
-                or np.random.uniform(0.0, 1.0) <= self.p_teacher_forcing
-            ):
-                teacher_forced_frame = decoder_inputs[
-                    mel_outputs.size(1) * self.n_frames_per_step_current
-                ]
-
-                to_concat = (teacher_forced_frame,)
-                decoder_input = torch.cat(to_concat, dim=1)
-            else:
-
-                # NOTE(zach): we may need to concat these as we go to ensure that
-                # it's easy to retrieve the last n_frames_per_step_init frames.
-                to_concat = (
-                    self.prenet(
-                        mel_outputs[:, -1, -1 * self.n_frames_per_step_current :]
-                    ),
-                )
-                decoder_input = torch.cat(to_concat, dim=1)
-            # NOTE(zach): When training with fp16_run == True, decoder_rnn seems to run into
-            # issues with NaNs in gradient, maybe due to vanishing gradients.
-            # Disable half-precision for this call to work around the issue.
-            with autocast(enabled=False):
-                mel_output, gate_output, attention_weights = self.decode(
-                    decoder_input, None
-                )
-            mel_outputs = torch.cat(
-                [
-                    mel_outputs,
-                    mel_output[
-                        :, 0 : self.n_mel_channels * self.n_frames_per_step_current
-                    ].unsqueeze(1),
-                ],
-                dim=1,
-            )
-            gate_outputs += [gate_output.squeeze()] * self.n_frames_per_step_current
-            alignments += [attention_weights]
-
-        mel_outputs, gate_outputs, alignments = self.parse_decoder_outputs(
-            mel_outputs, gate_outputs, alignments
-        )
-
-        return mel_outputs, gate_outputs, alignments
-
-    def inference(self, memory, memory_lengths):
-        """Decoder inference
-        PARAMS
-        ------
-        memory: Encoder outputs
-
-        RETURNS
-        -------
-        mel_outputs: mel outputs from the decoder
-        gate_outputs: gate outputs from the decoder
-        alignments: sequence of attention weights from the decoder
-        """
-        decoder_input = self.get_go_frame(memory)
-        self.initialize_decoder_states(
-            memory, mask=~get_mask_from_lengths(memory_lengths)
-        )
-
-        B = memory.size(0)
-        mel_outputs = torch.empty(
-            B, 0, self.n_frames_per_step_current * self.n_mel_channels
-        )
-        if self.cudnn_enabled:
-            mel_outputs = mel_outputs.cuda()
-        gate_outputs, alignments = [], []
-
-        mel_lengths = torch.zeros(
-            [memory.size(0)], dtype=torch.int32, device=memory.device
-        )
-        not_finished = torch.ones(
-            [memory.size(0)], dtype=torch.int32, device=memory.device
-        )
-
-        while True:
-            to_cat = (self.prenet(decoder_input),)
-            decoder_input = torch.cat(to_cat, dim=1)
-            mel_output, gate_output, alignment = self.decode(decoder_input, None)
-            mel_output = mel_output[
-                :, 0 : self.n_mel_channels * self.n_frames_per_step_current
-            ].unsqueeze(1)
-
-            mel_outputs = torch.cat([mel_outputs, mel_output], dim=1)
-            gate_outputs += [gate_output.squeeze()] * self.n_frames_per_step_current
-            alignments += [alignment]
-
-            dec = (
-                torch.le(torch.sigmoid(gate_output), self.gate_threshold)
-                .to(torch.int32)
-                .squeeze(1)
-            )
-
-            not_finished = not_finished * dec
-            mel_lengths += not_finished
-
-            if torch.sum(not_finished) == 0:
-                break
-            if mel_outputs.shape[1] == self.max_decoder_steps:
-                print("Warning! Reached max decoder steps")
-                break
-
-            decoder_input = mel_output[:, -1, -1 * self.n_mel_channels :]
-        mel_outputs, gate_outputs, alignments = self.parse_decoder_outputs(
-            mel_outputs, gate_outputs, alignments
-        )
-
-        return mel_outputs, gate_outputs, alignments, mel_lengths
-
-    def inference_noattention(self, memory, attention_map):
-        """Decoder inference
-        PARAMS
-        ------
-        memory: Encoder outputs
-
-        RETURNS
-        -------
-        mel_outputs: mel outputs from the decoder
-        gate_outputs: gate outputs from the decoder
-        alignments: sequence of attention weights from the decoder
-        """
-        decoder_input = self.get_go_frame(memory)
-
-        self.initialize_decoder_states(memory, mask=None)
-
-        B = memory.size(0)
-        mel_outputs = torch.empty(
-            B, 0, self.n_frames_per_step_current * self.n_mel_channels
-        )
-        if self.cudnn_enabled:
-            mel_outputs = mel_outputs.cuda()
-        gate_outputs, alignments = [], []
-        for i in range(len(attention_map)):
-
-            attention = attention_map[i]
-            decoder_input = self.prenet(decoder_input)
-            mel_output, gate_output, alignment = self.decode(decoder_input, attention)
-            mel_output = mel_output[
-                :, 0 : self.n_mel_channels * self.n_frames_per_step_current
-            ].unsqueeze(1)
-
-            mel_outputs = torch.cat([mel_outputs, mel_output], dim=1)
-            gate_outputs += [gate_output.squeeze()] * self.n_frames_per_step_current
-            alignments += [alignment]
-
-            decoder_input = mel_output[:, -1, -1 * self.n_mel_channels :]
-
-        mel_outputs, gate_outputs, alignments = self.parse_decoder_outputs(
-            mel_outputs, gate_outputs, alignments
-        )
-
-        return mel_outputs, gate_outputs, alignments
-
-    def inference_partial_tf(self, memory, decoder_inputs, tf_until_idx, device="cpu"):
-        """Decoder inference with teacher-forcing up until tf_until_idx
-        PARAMS
-        ------
-        memory: Encoder outputs
-        decoder_inputs: Decoder inputs for teacher forcing. i.e. mel-specs
-        memory_lengths: Encoder output lengths for attention masking.
-
-        RETURNS
-        -------
-        mel_outputs: mel outputs from the decoder
-        gate_outputs: gate outputs from the decoder
-        alignments: sequence of attention weights from the decoder
-        """
-        if device == "cuda" and self.cudnn_enabled:
-            decoder_inputs = decoder_inputs.cuda()
-
-        B = memory.size(0)
-        decoder_inputs = self.parse_decoder_inputs(decoder_inputs)
-        decoder_inputs = decoder_inputs.reshape(
-            -1, decoder_inputs.size(1), self.n_mel_channels
-        )
-        decoder_input = self.get_go_frame(memory).unsqueeze(0)
-        decoder_inputs = torch.cat((decoder_input, decoder_inputs), dim=0)
-        decoder_inputs = self.prenet(decoder_inputs)
-
-        self.initialize_decoder_states(memory, mask=None)
-
-        mel_outputs = torch.empty(
-            B, 0, self.n_frames_per_step_current * self.n_mel_channels
-        )
-        if device == "cuda" and self.cudnn_enabled:
-            mel_outputs = mel_outputs.cuda()
-        gate_outputs, alignments = [], []
-
-        while True:
-            if mel_outputs.size(1) < tf_until_idx:
-                teacher_forced_frame = decoder_inputs[
-                    mel_outputs.size(1) * self.n_frames_per_step_current
-                ]
-
-                to_concat = (teacher_forced_frame,)
-                decoder_input = torch.cat(to_concat, dim=1)
-            else:
-
-                to_concat = (
-                    self.prenet(mel_outputs[:, -1, -1 * self.n_mel_channels :]),
-                )
-                decoder_input = torch.cat(to_concat, dim=1)
-            mel_output, gate_output, attention_weights = self.decode(
-                decoder_input, None
-            )
-            mel_outputs = torch.cat(
-                [
-                    mel_outputs,
-                    mel_output[
-                        :, 0 : self.n_mel_channels * self.n_frames_per_step_current
-                    ].unsqueeze(1),
-                ],
-                dim=1,
-            )
-            gate_outputs += [gate_output.squeeze()] * self.n_frames_per_step_current
-            alignments += [attention_weights]
-            if torch.sigmoid(gate_output.data) > self.gate_threshold:
-                break
-            elif mel_outputs.size(1) == self.max_decoder_steps:
-                print("Warning! Reached max decoder steps")
-                break
-
-        mel_outputs, gate_outputs, alignments = self.parse_decoder_outputs(
-            mel_outputs, gate_outputs, alignments
-        )
-
-        return mel_outputs, gate_outputs, alignments
-
-
-class Prenet(nn.Module):
-    def __init__(self, in_dim, sizes):
-        super().__init__()
-        in_sizes = [in_dim] + sizes[:-1]
-        self.layers = nn.ModuleList(
-            [
-                LinearNorm(in_size, out_size, bias=False)
-                for (in_size, out_size) in zip(in_sizes, sizes)
-            ]
-        )
-        self.dropout_rate = 0.5
-
-    def forward(self, x):
-        for linear in self.layers:
-            x = F.dropout(F.relu(linear(x)), p=self.dropout_rate, training=True)
-        return x
-
-
-class Postnet(nn.Module):
-    """Postnet
-    - Five 1-d convolution with 512 channels and kernel size 5
-    """
-
-    def __init__(self, hparams):
-        super(Postnet, self).__init__()
-        self.dropout_rate = 0.5
-        self.convolutions = nn.ModuleList()
-
-        self.convolutions.append(
-            nn.Sequential(
-                Conv1d(
-                    hparams.n_mel_channels,
-                    hparams.postnet_embedding_dim,
-                    kernel_size=hparams.postnet_kernel_size,
-                    stride=1,
-                    padding=int((hparams.postnet_kernel_size - 1) / 2),
-                    dilation=1,
-                    w_init_gain="tanh",
-                ),
-                nn.BatchNorm1d(hparams.postnet_embedding_dim),
-            )
-        )
-
-        for i in range(1, hparams.postnet_n_convolutions - 1):
-            self.convolutions.append(
-                nn.Sequential(
-                    Conv1d(
-                        hparams.postnet_embedding_dim,
-                        hparams.postnet_embedding_dim,
-                        kernel_size=hparams.postnet_kernel_size,
-                        stride=1,
-                        padding=int((hparams.postnet_kernel_size - 1) / 2),
-                        dilation=1,
-                        w_init_gain="tanh",
-                    ),
-                    nn.BatchNorm1d(hparams.postnet_embedding_dim),
-                )
-            )
-
-        self.convolutions.append(
-            nn.Sequential(
-                Conv1d(
-                    hparams.postnet_embedding_dim,
-                    hparams.n_mel_channels,
-                    kernel_size=hparams.postnet_kernel_size,
-                    stride=1,
-                    padding=int((hparams.postnet_kernel_size - 1) / 2),
-                    dilation=1,
-                    w_init_gain="linear",
-                ),
-                nn.BatchNorm1d(hparams.n_mel_channels),
-            )
-        )
-
-    def forward(self, x):
-        for i, conv in enumerate(self.convolutions):
-            if i == len(self.convolutions) - 1:
-                x = F.dropout(conv(x), self.dropout_rate, self.training)
-            else:
-                x = F.dropout(torch.tanh(conv(x)), self.dropout_rate, self.training)
-
-        return x
-
-
-class Encoder(nn.Module):
-    """Encoder module:
-    - Three 1-d convolution banks
-    - Bidirectional LSTM
-    """
-
-    def __init__(self, hparams):
-        super().__init__()
-
-        convolutions = []
-        for _ in range(hparams.encoder_n_convolutions):
-            conv_layer = nn.Sequential(
-                Conv1d(
-                    hparams.encoder_embedding_dim,
-                    hparams.encoder_embedding_dim,
-                    kernel_size=hparams.encoder_kernel_size,
-                    stride=1,
-                    padding=int((hparams.encoder_kernel_size - 1) / 2),
-                    dilation=1,
-                    w_init_gain="relu",
-                ),
-                nn.BatchNorm1d(hparams.encoder_embedding_dim),
-            )
-            convolutions.append(conv_layer)
-        self.convolutions = nn.ModuleList(convolutions)
-        self.dropout_rate = 0.5
-
-        self.lstm = nn.LSTM(
-            hparams.encoder_embedding_dim,
-            int(hparams.encoder_embedding_dim / 2),
-            1,
-            batch_first=True,
-            bidirectional=True,
-        )
-
-    def forward(self, x, input_lengths):
-        if x.size()[0] > 1:
-            x_embedded = []
-            for b_ind in range(x.size()[0]):  # TODO: Speed up
-                curr_x = x[b_ind : b_ind + 1, :, : input_lengths[b_ind]].clone()
-                for conv in self.convolutions:
-                    curr_x = F.dropout(
-                        F.relu(conv(curr_x)), self.dropout_rate, self.training
-                    )
-                x_embedded.append(curr_x[0].transpose(0, 1))
-            x = torch.nn.utils.rnn.pad_sequence(x_embedded, batch_first=True)
-        else:
-            for conv in self.convolutions:
-                x = F.dropout(F.relu(conv(x)), self.dropout_rate, self.training)
-            x = x.transpose(1, 2)
-
-        # pytorch tensor are not reversible, hence the conversion
-        input_lengths = input_lengths.cpu().numpy()
-        x = nn.utils.rnn.pack_padded_sequence(
-            x, input_lengths, batch_first=True, enforce_sorted=False
-        )
-
-        self.lstm.flatten_parameters()
-        outputs, _ = self.lstm(x)
-
-        outputs, _ = nn.utils.rnn.pad_packed_sequence(outputs, batch_first=True)
-        return outputs
-
-    def inference(self, x, input_lengths):
-        device = x.device
-        for conv in self.convolutions:
-            x = F.dropout(F.relu(conv(x)), self.dropout_rate, self.training)
-
-        x = x.transpose(1, 2)
-
-        input_lengths = input_lengths.cpu()
-        x = nn.utils.rnn.pack_padded_sequence(
-            x, input_lengths, batch_first=True, enforce_sorted=False
-        )
-
-        outputs, _ = self.lstm(x)
-
-        outputs, _ = nn.utils.rnn.pad_packed_sequence(outputs, batch_first=True)
-
-        return outputs
-
-
+from .base import TTSModel
 from ..vendor.tfcompat.hparam import HParams
 from .base import DEFAULTS as MODEL_DEFAULTS
+from .components.decoders.tacotron2 import Decoder, DecoderForwardIsInfer
+from .components.encoders.tacotron2 import Encoder, EncoderForwardIsInfer
+from .components.postnet import Postnet
+from .components.zero_network import ZeroNetwork
 
 DEFAULTS = HParams(
     symbols_embedding_dim=512,
@@ -756,12 +83,6 @@ DEFAULTS = HParams(
 config = DEFAULTS.values()
 config.update(MODEL_DEFAULTS.values())
 DEFAULTS = HParams(**config)
-
-# NOTE (Sam): this is a hack that enables torchscipt compilation with n_speakers = 1 and has_speaker_embedding=True
-class ZeroNetwork(nn.Module):
-    def forward(self, x):
-
-        return torch.zeros_like(x)
 
 
 class Tacotron2(TTSModel):
@@ -1015,17 +336,6 @@ class Tacotron2(TTSModel):
         )
 
         return output
-
-
-# NOTE (Sam): for torchscipt compilation
-class EncoderForwardIsInfer(Encoder):
-    def forward(self, x, input_lengths):
-        return self.inference(x, input_lengths)
-
-
-class DecoderForwardIsInfer(Decoder):
-    def forward(self, memory, memory_lengths):
-        return self.inference(memory, memory_lengths)
 
 
 class Tacotron2ForwardIsInfer(Tacotron2):
